+++ jackson/src/test/scala/com/twitter/finatra/json/tests/internal/caseclass/validation/CaseClassValidationTest.scala
@@ -6,12 +6,17 @@ import com.twitter.finatra.json.internal.caseclass.exceptions.{
   CaseClassMappingException,
   CaseClassValidationException
 }
+import com.twitter.finatra.json.internal.caseclass.validation.validators.{
+  MinInternal,
+  NotEmptyInternal
+}
 import com.twitter.finatra.json.tests.internal.CarMake
 import com.twitter.finatra.json.tests.internal.caseclass.validation.domain.{Address, Car, Person}
-import com.twitter.finatra.validation.ErrorCode
+import com.twitter.finatra.validation.{ErrorCode, MethodValidation}
 import com.twitter.finatra.validation.ValidationResult.Invalid
 import com.twitter.inject.Test
 import org.joda.time.DateTime
+import org.scalatest.OptionValues._
 
 class CaseClassValidationTest extends Test {
 
@@ -41,19 +46,14 @@ class CaseClassValidationTest extends Test {
       parseCar(baseCar.copy(id = 2, year = 1910))
     }
 
-    parseError should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.leaf("year"),
-            Invalid(
-              "[1910] is not greater than or equal to 2000",
-              ErrorCode.ValueTooSmall(2000, 1910)
-            )
-          )
-        )
-      )
-    )
+    parseError.errors.size shouldEqual 1
+
+    val error = parseError.errors.head
+    error.path shouldEqual PropertyPath.leaf("year")
+    error.reason.code shouldEqual ErrorCode.ValueTooSmall(2000, 1910)
+    error.reason.message shouldEqual "[1910] is not greater than or equal to 2000"
+    error.reason.annotation shouldBe defined
+    error.reason.annotation.value shouldBe a [MinInternal]
   }
 
   test("class and field level validations#nested failed validations") {
@@ -76,20 +76,22 @@ class CaseClassValidationTest extends Test {
       parseCar(car)
     }
 
-    parseError should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.leaf("city").withParent("address").withParent("owners"),
-            Invalid("cannot be empty", ErrorCode.ValueCannotBeEmpty)
-          ),
-          CaseClassValidationException(
-            PropertyPath.leaf("street").withParent("address").withParent("owners"),
-            Invalid("cannot be empty", ErrorCode.ValueCannotBeEmpty)
-          )
-        )
-      )
-    )
+    parseError.errors.size shouldEqual 2
+    val errors = parseError.errors
+
+    errors.head.path shouldEqual
+      PropertyPath.leaf("city").withParent("address").withParent("owners")
+    errors.head.reason.code shouldEqual ErrorCode.ValueCannotBeEmpty
+    errors.head.reason.message shouldEqual "cannot be empty"
+    errors.head.reason.annotation shouldBe defined
+    errors.head.reason.annotation.value shouldBe a [NotEmptyInternal]
+
+    errors(1).path shouldEqual
+      PropertyPath.leaf("street").withParent("address").withParent("owners")
+    errors(1).reason.code shouldEqual ErrorCode.ValueCannotBeEmpty
+    errors(1).reason.message shouldEqual "cannot be empty"
+    errors(1).reason.annotation shouldBe defined
+    errors(1).reason.annotation.value shouldBe a [NotEmptyInternal]
   }
 
   test("class and field level validations#nested method validations") {
@@ -106,73 +108,84 @@ class CaseClassValidationTest extends Test {
       parseCar(car)
     }
 
-    parseError should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.leaf("address").withParent("owners"),
-            Invalid("state must be one of [CA, MD, WI]")
-          )
-        )
-      )
-    )
+    parseError.errors.size shouldEqual 1
+    val errors = parseError.errors
 
-    parseError.errors.map(_.getMessage) should equal(
+    errors.head.path shouldEqual
+      PropertyPath.leaf("address").withParent("owners")
+    errors.head.reason.code shouldEqual ErrorCode.Unknown
+    errors.head.reason.message shouldEqual "state must be one of [CA, MD, WI]"
+    errors.head.reason.annotation shouldBe defined
+    errors.head.reason.annotation.value shouldBe a [MethodValidation]
+
+    errors.map(_.getMessage) should equal(
       Seq("owners.address: state must be one of [CA, MD, WI]")
     )
   }
 
   test("class and field level validations#end before start") {
-    intercept[CaseClassMappingException] {
+    val parseError = intercept[CaseClassMappingException] {
       parseCar(
         baseCar.copy(ownershipStart = baseCar.ownershipEnd, ownershipEnd = baseCar.ownershipStart)
       )
-    } should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.empty,
-            Invalid(
-              "ownershipEnd [2015-04-09T05:17:15.000Z] must be after ownershipStart [2015-04-09T05:18:15.000Z]"
-            )
-          )
-        )
-      )
-    )
+    }
+
+    parseError.errors.size shouldEqual 1
+    val errors = parseError.errors
+
+    errors.head.path shouldEqual PropertyPath.leaf("ownershipEnd")
+    errors.head.reason.code shouldEqual ErrorCode.Unknown
+    errors.head.reason.message shouldEqual "ownershipEnd [2015-04-09T05:17:15.000Z] must be after ownershipStart [2015-04-09T05:18:15.000Z]"
+    errors.head.reason.annotation shouldBe defined
+    errors.head.reason.annotation.value shouldBe a [MethodValidation]
   }
 
   test("class and field level validations#optional end before start") {
-    intercept[CaseClassMappingException] {
+    val parseError = intercept[CaseClassMappingException] {
       parseCar(
         baseCar.copy(warrantyStart = baseCar.warrantyEnd, warrantyEnd = baseCar.warrantyStart)
       )
-    } should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.empty,
-            Invalid(
-              "warrantyEnd [2015-04-09T05:17:15.000Z] must be after warrantyStart [2015-04-09T06:17:15.000Z]"
-            )
-          )
-        )
-      )
-    )
+    }
+
+    parseError.errors.size shouldEqual 2
+    val errors = parseError.errors
+
+    errors.head.path shouldEqual PropertyPath.leaf("warrantyEnd")
+    errors.head.reason.code shouldEqual ErrorCode.Unknown
+    errors.head.reason.message shouldEqual
+      "warrantyEnd [2015-04-09T05:17:15.000Z] must be after warrantyStart [2015-04-09T06:17:15.000Z]"
+    errors.head.reason.annotation shouldBe defined
+    errors.head.reason.annotation.value shouldBe a [MethodValidation]
+
+    errors(1).path shouldEqual PropertyPath.leaf("warrantyStart")
+    errors(1).reason.code shouldEqual ErrorCode.Unknown
+    errors(1).reason.message shouldEqual
+      "warrantyEnd [2015-04-09T05:17:15.000Z] must be after warrantyStart [2015-04-09T06:17:15.000Z]"
+    errors(1).reason.annotation shouldBe defined
+    errors(1).reason.annotation.value shouldBe a [MethodValidation]
   }
 
   test("class and field level validations#no start with end") {
-    intercept[CaseClassMappingException] {
+    val parseError = intercept[CaseClassMappingException] {
       parseCar(baseCar.copy(warrantyStart = None, warrantyEnd = baseCar.warrantyEnd))
-    } should equal(
-      CaseClassMappingException(
-        Set(
-          CaseClassValidationException(
-            PropertyPath.empty,
-            Invalid("both warrantyStart and warrantyEnd are required for a valid range")
-          )
-        )
-      )
-    )
+    }
+
+    parseError.errors.size shouldEqual 2
+    val errors = parseError.errors
+
+    errors.head.path shouldEqual PropertyPath.leaf("warrantyEnd")
+    errors.head.reason.code shouldEqual ErrorCode.Unknown
+    errors.head.reason.message shouldEqual
+      "both warrantyStart and warrantyEnd are required for a valid range"
+    errors.head.reason.annotation shouldBe defined
+    errors.head.reason.annotation.value shouldBe a [MethodValidation]
+
+    errors(1).path shouldEqual PropertyPath.leaf("warrantyStart")
+    errors(1).reason.code shouldEqual ErrorCode.Unknown
+    errors(1).reason.message shouldEqual
+      "both warrantyStart and warrantyEnd are required for a valid range"
+    errors(1).reason.annotation shouldBe defined
+    errors(1).reason.annotation.value shouldBe a [MethodValidation]
   }
 
   test("class and field level validations#start with end") {